refactor: split scheduler runtime and add task ttl cleanup
