+++ src/config/task_scheduler/maintenance.rs
@@ -1,8 +1,7 @@
-use std::{env, time::Duration};
+use std::time::Duration;
 
 use tokio::time::MissedTickBehavior;
 use tracing::warn;
-use uuid::Uuid;
 
 use crate::{
     domain::{
@@ -16,34 +15,6 @@ const DEFAULT_WORKER_HEARTBEAT_INTERVAL_SECONDS: u64 = 15;
 const DEFAULT_TIMEOUT_SWEEP_INTERVAL_SECONDS: u64 = 30;
 const DEFAULT_WORKER_STALE_AFTER_SECONDS: i64 = 30 * 60;
 const DEFAULT_TIMEOUT_SWEEP_LIMIT: usize = 100;
-
-pub fn default_task_scheduler_worker_id() -> String {
-    resolve_task_scheduler_worker_id(
-        env::var("TASK_SCHEDULER_WORKER_ID").ok(),
-        env::var("HOSTNAME").ok(),
-    )
-}
-
-fn resolve_task_scheduler_worker_id(
-    explicit_worker_id: Option<String>,
-    hostname: Option<String>,
-) -> String {
-    if let Some(worker_id) = explicit_worker_id
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-    {
-        return worker_id;
-    }
-
-    if let Some(hostname) = hostname
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-    {
-        return format!("task-scheduler-{hostname}");
-    }
-
-    format!("task-scheduler-{}", Uuid::new_v4())
-}
 
 #[derive(Clone, Debug)]
 pub struct TaskSchedulerWorkerConfig {
@@ -135,33 +106,4 @@ where
     });
 
     BackgroundTaskHandle::new("task-scheduler-maintenance", shutdown_tx, join_handle)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::resolve_task_scheduler_worker_id;
-
-    #[test]
-    fn default_worker_id_prefers_explicit_env() {
-        let worker_id = resolve_task_scheduler_worker_id(
-            Some("scheduler-a".to_string()),
-            Some("container-1".to_string()),
-        );
-
-        assert_eq!(worker_id, "scheduler-a");
-    }
-
-    #[test]
-    fn default_worker_id_falls_back_to_hostname() {
-        let worker_id = resolve_task_scheduler_worker_id(None, Some("container-2".to_string()));
-
-        assert_eq!(worker_id, "task-scheduler-container-2");
-    }
-
-    #[test]
-    fn default_worker_id_falls_back_to_uuid_when_env_missing() {
-        let worker_id = resolve_task_scheduler_worker_id(None, None);
-
-        assert!(worker_id.starts_with("task-scheduler-"));
-    }
 }
