+++ src/config/task_scheduler/worker.rs
@@ -1,47 +1,20 @@
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
 use futures::future::join_all;
-use tokio::sync::Mutex;
-use tokio::sync::{watch, OwnedSemaphorePermit, Semaphore};
+use tokio::sync::{watch, Mutex, OwnedSemaphorePermit, Semaphore};
 use tokio::time::MissedTickBehavior;
-use tracing::{info, warn};
+use tracing::warn;
 
-use crate::domain::identity::Clock;
-use crate::BackgroundTaskHandle;
-
-use super::{
-    BoxFuture, FailTaskInput, ScheduledTask, TaskRepository, TaskSchedulerError,
-    TaskSchedulerService, TaskWorkerRepository,
+use crate::{
+    domain::{
+        identity::Clock,
+        task_scheduler::{
+            FailTaskInput, ScheduledTask, SharedTaskHandler, TaskRepository, TaskRunOutcome,
+            TaskSchedulerError, TaskSchedulerService, TaskWorkerConfig, TaskWorkerRepository,
+        },
+    },
+    BackgroundTaskHandle,
 };
-
-#[derive(Clone, Debug)]
-pub struct TaskWorkerConfig {
-    pub is_leader: bool,
-    pub lease_duration_seconds: i64,
-    pub heartbeat_interval: Duration,
-    pub idle_poll_interval: Duration,
-    pub max_concurrency: usize,
-}
-
-#[derive(Debug)]
-pub enum TaskRunOutcome {
-    Completed {
-        checkpoint: Option<serde_json::Value>,
-    },
-    Failed {
-        checkpoint: Option<serde_json::Value>,
-        error_message: String,
-        retryable: bool,
-    },
-}
-
-pub trait TaskHandler: Send + Sync + 'static {
-    fn task_type(&self) -> &'static str;
-
-    fn run(&self, task: ScheduledTask) -> BoxFuture<TaskRunOutcome>;
-}
-
-pub type SharedTaskHandler = Arc<dyn TaskHandler>;
 
 struct TaskHandlerRegistry {
     enabled_task_types: Vec<String>,
@@ -306,7 +279,7 @@ where
         runtime_state.clone(),
     );
 
-    info!(
+    tracing::info!(
         task_id = %task.id,
         task_type = %task.task_type,
         worker_id = %worker_id,
