+++ src/domain/task_scheduler/mod.rs
@@ -1,7 +1,7 @@
 mod model;
 mod ports;
-mod runner;
 mod service;
+mod worker;
 
 pub use model::{
     NewTask, RetryStrategy, ScheduledTask, TaskCheckpointRequest, TaskClaimRequest,
@@ -10,7 +10,5 @@ pub use model::{
     TaskWorker,
 };
 pub use ports::{BoxFuture, TaskRepository, TaskWorkerRepository};
-pub use runner::{
-    spawn_task_worker, SharedTaskHandler, TaskHandler, TaskRunOutcome, TaskWorkerConfig,
-};
 pub use service::{FailTaskInput, ResultTaskHandler, TaskSchedulerService};
+pub use worker::{SharedTaskHandler, TaskHandler, TaskRunOutcome, TaskWorkerConfig};
